Restart the projection subsystem
Add an endpoint to the projections controller that allows restarting the
projection subsystem.

Handle the restart message in the projection subsystem, which publishes
a restart subcomponents message.

Add a new core projection state, Suspended, that unsubscribes the
projection's event readers when entered.

Coordinate stopping projections in ProjectionCoreCoordinator:

Try and coordinate the projection shutdown a bit better.
1. Publish the StopCore message to start suspending the projections
2. Track which projections should be suspended when suspending them, and
listen for the projection's "Suspended" message before considering the
ProjectionCoreSystem as fully stopped.
3. When ProjectionCoreService is stopped, stop the readers for that
queue.
4. The subscriptions and event readers should have been cleaned up in
the EventReaderCoreService with the projections being suspended in step 2.

ProjectionManager stops itself, clears its projections and starts again.